admin tpin verification

+++ src/components/admin/JobDescriptionAdmin.js
@@ -67,6 +67,13 @@ const JobDescriptionAdmin = () => {
   }, []);
 
   const handleDeleteJob = async () => {
+    // verify admin using admin tpin
+    const TPIN = window.prompt("Enter Admin TPIN to proceed 👽", "");
+    if (TPIN !== process.env.REACT_APP_ADMIN_TPIN) {
+      toast(ToastConfig("Error", "Invalid TPIN", "error"));
+      return;
+    }
+
     try {
       setBtnLoading(true);
       const data = await deleteJobAdmin(jobId);
